scraper/requirements.txt
playwright==1.40.0
pytesseract==0.3.10
Pillow==10.1.0
supabase==2.3.0
python-dotenv==1.0.0
